requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
argcomplete==1.8.1
async-generator==1.10
attrs==22.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
chrome-gnome-shell==0.0.0
chromedriver-autoinstaller==0.6.2
click==8.1.6
cmake==3.27.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
configparser==6.0.1
contourpy==1.1.0
coverage==7.4.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
duplicity==0.8.21
exceptiongroup==1.1.1
fasteners==0.14.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.41.1
future==0.18.2
gdown==4.7.1
h11==0.14.0
html-table-parser==0.1.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.4
ijson==3.2.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightgbm==4.0.0
lit==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.7.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
openai==1.14.2
opencv-python==4.8.0.74
osmium==3.7.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.0.0
protobuf==3.12.4
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py4j==0.10.9.7
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
pyspark==3.5.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
sdkmanager==0.5.1
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.8.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
speedtest-cli==2.1.3
ssh-import-id==5.11
sympy==1.12
systemd-python==234
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-cleaner==1.1.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.0.142
unattended-upgrades==0.1
urllib3==1.26.17
urwid==2.1.2
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webdriver-manager==3.8.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
